right a wrong (renamed the dir, should only rename file)

+++ cmd.sh
@@ -1,3 +1,5 @@
 for f in `ls | grep -v '[.]'`; do
-	mv $f ${f}.md
+	if [ -f f ]; then 
+		mv $f ${f}.md
+	fi
 done
